refactor: rename to Infinity Agent (name, namespace, applicationId, packages)
