+++ app/build.gradle.kts
@@ -4,11 +4,11 @@ plugins {
 }
 
 android {
-    namespace = "com.infinityempire.manus"
+    namespace = "com.infinityempire.infinityagent"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.infinityempire.manus"
+        applicationId = "com.infinityempire.infinityagent"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
