+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Manus Agent</string>
+    <string name="app_name">Infinity Agent</string>
 </resources>
